requirements.txt
pdfplumber>=0.10.0
google-genai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
regex>=2023.0.0
